完成部分管理界面设计

+++ PornTube/src/main/java/com/buguagaoshu/porntube/utils/PageUtils.java
@@ -14,19 +14,19 @@ public class PageUtils implements Serializable {
     /**
      * 总记录数
      */
-    private int totalCount;
+    private long totalCount;
     /**
      * 每页记录数
      */
-    private int pageSize;
+    private long pageSize;
     /**
      * 总页数
      */
-    private int totalPage;
+    private long totalPage;
     /**
      * 当前页数
      */
-    private int currPage;
+    private long currPage;
     /**
      * 列表数据
      */
@@ -39,7 +39,7 @@ public class PageUtils implements Serializable {
      * @param pageSize    每页记录数
      * @param currPage    当前页数
      */
-    public PageUtils(List<?> list, int totalCount, int pageSize, int currPage) {
+    public PageUtils(List<?> list, long totalCount, long pageSize, long currPage) {
         this.list = list;
         this.totalCount = totalCount;
         this.pageSize = pageSize;
@@ -52,13 +52,13 @@ public class PageUtils implements Serializable {
      */
     public PageUtils(IPage<?> page) {
         this.list = page.getRecords();
-        this.totalCount = (int)page.getTotal();
-        this.pageSize = (int)page.getSize();
-        this.currPage = (int)page.getCurrent();
-        this.totalPage = (int)page.getPages();
+        this.totalCount = page.getTotal();
+        this.pageSize = page.getSize();
+        this.currPage = page.getCurrent();
+        this.totalPage = page.getPages();
     }
 
-    public int getTotalCount() {
+    public long getTotalCount() {
         return totalCount;
     }
 
@@ -66,7 +66,7 @@ public class PageUtils implements Serializable {
         this.totalCount = totalCount;
     }
 
-    public int getPageSize() {
+    public long getPageSize() {
         return pageSize;
     }
 
@@ -74,7 +74,7 @@ public class PageUtils implements Serializable {
         this.pageSize = pageSize;
     }
 
-    public int getTotalPage() {
+    public long getTotalPage() {
         return totalPage;
     }
 
@@ -82,7 +82,7 @@ public class PageUtils implements Serializable {
         this.totalPage = totalPage;
     }
 
-    public int getCurrPage() {
+    public long getCurrPage() {
         return currPage;
     }
